feat: enhance LunaVDB with improved add/remove methods and update types for better performance

+++ src/engine/types.rs
@@ -13,7 +13,7 @@ const BUCKET_SIZE: usize = 32;
 
 pub type Embedding = Vec<f32>;
 
-pub type Tree = KdTree<f32, u64, 1024, BUCKET_SIZE, u32>;
+pub type Tree = KdTree<f32, u64, 1024, BUCKET_SIZE, u16>;
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct Index {
